+++ je_auto_control/utils/text_unicode/__init__.py
@@ -1,0 +1,6 @@
+"""Type arbitrary Unicode (emoji / CJK / accented) via the clipboard."""
+from je_auto_control.utils.text_unicode.text_unicode import (
+    plan_paste, type_unicode, unicode_code_units,
+)
+
+__all__ = ["plan_paste", "type_unicode", "unicode_code_units"]